feat(Modules): re structured modules to split files for clarity
