+++ Runtime/CVRFury/SubComponents/Modules/CVRFuryModulesArmatureLink.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 130f9e0bfcc4b7d4486a2ebd34e590bd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
