+++ Runtime/CVRFury/SubComponents/Modules/CVRFuryModulesBlockBlinkSetter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3b7e5f10adb369f4aa37223790032d1f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
