+++ Runtime/CVRFury/SubComponents/Modules/CVRFuryModulesBlockVisemesSetter.cs
@@ -1,0 +1,19 @@
+using System;
+using System.Collections.Generic;
+using UnityEngine;
+using uk.novavoidhowl.dev.vrcstub;
+using uk.novavoidhowl.dev.cvrfury.supporting_classes.runtime;
+using uk.novavoidhowl.dev.cvrfury.packagecore;
+using System.Reflection;
+
+namespace uk.novavoidhowl.dev.cvrfury.runtime
+{
+  [Serializable]
+  public class blockVisemesSetter : StateSetter
+  {
+    public blockVisemesSetter()
+      : base("blockVisemesSetter") { }
+
+    public bool blockVisemes;
+  }
+}
