+++ Runtime/CVRFury/SubComponents/Modules/CVRFuryModulesBlockVisemesSetter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fb160de46a99f144591bc6c1a0ef584e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
