+++ Runtime/CVRFury/SubComponents/Modules/CVRFuryModulesCore.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 135b1b72232a2094b865c17e3beb58cf
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
