+++ Runtime/CVRFury/SubComponents/Modules/CVRFuryModulesCustomLink.cs
@@ -1,0 +1,19 @@
+using System;
+using System.Collections.Generic;
+using UnityEngine;
+using uk.novavoidhowl.dev.vrcstub;
+using uk.novavoidhowl.dev.cvrfury.supporting_classes.runtime;
+using uk.novavoidhowl.dev.cvrfury.packagecore;
+using System.Reflection;
+
+namespace uk.novavoidhowl.dev.cvrfury.runtime
+{
+  [Serializable]
+  public class customLink : CVRFuryModuleLinker
+  {
+    public customLink()
+      : base("CustomLink") { }
+
+    public GameObject targetArmatureObject;
+  }
+}
