+++ Runtime/CVRFury/SubComponents/Modules/CVRFuryModulesCustomLink.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fa60cb85d5b443a4192d8a0f6b61c15d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
