+++ Runtime/CVRFury/SubComponents/Modules/CVRFuryModulesDeleteOnUpload.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4c21e9b25dc5d8c4c934ee483f22a8c4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
