+++ Runtime/CVRFury/SubComponents/Modules/CVRFuryModulesFullController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 92c93ec43ecef8d4690c21d586e5fff7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
