+++ Runtime/CVRFury/SubComponents/Modules/CVRFuryModulesMenuLink.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c78e587f20f0df149a6cd5699942f930
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
