+++ Runtime/CVRFury/SubComponents/Modules/CVRFuryModulesObjectDefaultBlendShapeSetter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 31fa0e578eacd4646809c596f0feb70b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
