+++ Runtime/CVRFury/SubComponents/Modules/CVRFuryModulesObjectDefaultMaterialSetter.cs
@@ -1,0 +1,26 @@
+using System;
+using System.Collections.Generic;
+using UnityEngine;
+using uk.novavoidhowl.dev.vrcstub;
+using uk.novavoidhowl.dev.cvrfury.supporting_classes.runtime;
+using uk.novavoidhowl.dev.cvrfury.packagecore;
+using System.Reflection;
+
+namespace uk.novavoidhowl.dev.cvrfury.runtime
+{
+  [Serializable]
+  public class objectDefaultMaterialSetter : StateSetter
+  {
+    public objectDefaultMaterialSetter()
+      : base("objectDefaultMaterialSetter") { }
+
+    // material to set as default
+    public Material defaultMaterial;
+
+    // renderer to set the default material on
+    public Renderer renderer;
+
+    // renderer index to set the default material on
+    public int rendererIndex;
+  }
+}
