+++ Runtime/CVRFury/SubComponents/Modules/CVRFuryModulesObjectDefaultMaterialSetter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bef5a4719237f204e8bcee1ec8276fc9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
