+++ Runtime/CVRFury/SubComponents/Modules/CVRFuryModulesObjectFxFloatSetter.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 51fdb62f7fb0a1a4095753e0bd1731ea
+guid: 42585426ae5d77c4e8dd51b79541b3af
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
