+++ Runtime/CVRFury/SubComponents/Modules/CVRFuryModulesObjectScaleSetter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 45d731c2f98bdfe41a745136f5769f58
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
