+++ Runtime/CVRFury/SubComponents/Modules/CVRFuryModulesObjectStateSetter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c085559a6129f954b9568f771280e2ec
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
