+++ Runtime/CVRFury/SubComponents/Modules/CVRFuryModulesParameterStreamLink.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e5ca1eece89fab644bedf3f20eede5cc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
