+++ Runtime/CVRFury/SubComponents/Modules/CVRFuryModulesParametersLink.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9580c165a12c159469ada2ebe68778b2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
