+++ Runtime/CVRFury/SubComponents/Modules/CVRFuryModulesPrefabDependency.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7e8eda0d231750140b131cd069bd753d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
