+++ Runtime/CVRFury/SubComponents/Modules/CVRFuryModulesShowInFirstPerson.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cc1bcca9c555f9440966e097cf61ae37
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
