+++ package.json
@@ -22,5 +22,5 @@
   },
   "unity": "2021.3",
   "unityRelease": "23f1",
-  "version": "0.157.1-alpha"
+  "version": "0.158.0-alpha"
 }
